renamed: #67 フックの種類が多かったので章ごとにディレクトリを分けて探しやすくする
